Section 1
Add a theme view
Users can select a color theme to help differentiate their daily scrums. In this section, you’ll create a theme view that displays the components of a theme.

In the next section, the app will repeatedly use this theme view to display all the themes in a list.