Section 2 - Add a theme picker
In this section, you’ll create a custom interactive view for your users to choose the color theme of a meeting view. Using the theme view, you’ll list all your app’s themes in a picker control. Then, you’ll store the value of the selected theme in a variable that you bind to the parent view. In this way, the theme view communicates a theme selection up to the parent view.